Update index.css
I added some simple animation for one of the stretch goals. The "Learn" buttons on the about page change the background color to black and text color to white the cursor hovers over them. The cursor also changes to a pointer. The navigation items add an underline when the pointer hovers over thos elements as well.